Separate plugin by insert or load
